Get Planting to Work
NEEDED:
Basic Tractor Prefab
Planting Area Prefab

This is set to dynamically make planter spots within the planting area. Set the size of your planting area to do so. The tractor will highlighted where the crops will be planted and you must place space to plant crops. The number of the crops to be planted is inherited from the size of the planting box.

See my Scene to test and ask if questions